Refactor to fix codeclimate issues
